Move email route from index to email.route.js
